# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Barbearia Rocky Amaral
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
}

export interface CreateAppointmentBody {
  clientName: string;
  clientPhone: string;
  serviceId: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  notes?: string;
}

export interface Appointment {
  id: number;
  clientName: string;
  clientPhone: string;
  serviceId: number;
  serviceName: string;
  date: string;
  time: string;
  notes?: string;
  status: string;
  whatsappLink: string;
  createdAt: string;
}

export interface AvailableTimes {
  date: string;
  times: string[];
}

export interface ErrorResponse {
  error: string;
}

export type GetAvailableTimesParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
};