# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Barbearia Rocky Amaral
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all barber services
 */
export const ListServicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  durationMinutes: zod.number(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * @summary Create a new appointment
 */
export const CreateAppointmentBody = zod.object({
  clientName: zod.string(),
  clientPhone: zod.string(),
  serviceId: zod.number(),
  date: zod.string().describe("Date in YYYY-MM-DD format"),
  time: zod.string().describe("Time in HH:MM format"),
  notes: zod.string().optional(),
});

/**
 * @summary Get available time slots for a date
 */
export const GetAvailableTimesQueryParams = zod.object({
  date: zod.coerce.string(),
});

export const GetAvailableTimesResponse = zod.object({
  date: zod.string(),
  times: zod.array(zod.string()),
});