# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Barbearia Rocky Amaral
 * OpenAPI spec version: 0.1.0
 */

export interface Appointment {
  id: number;
  clientName: string;
  clientPhone: string;
  serviceId: number;
  serviceName: string;
  date: string;
  time: string;
  notes?: string;
  status: string;
  whatsappLink: string;
  createdAt: string;
}