# lib/api-zod/src/generated/types/availableTimes.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Barbearia Rocky Amaral
 * OpenAPI spec version: 0.1.0
 */

export interface AvailableTimes {
  date: string;
  times: string[];
}