# lib/api-zod/src/generated/types/createAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Barbearia Rocky Amaral
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppointmentBody {
  clientName: string;
  clientPhone: string;
  serviceId: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  notes?: string;
}